Add as funcoes de calculo(Age,Attack,Exp)

+++ src/knightsApi/knights.Core/Entities/Attributes.cs
@@ -3,11 +3,11 @@ namespace Knights.Core.Entities
     public class Attributes
     {
         public Guid Id { get; set; }
-        public double Strength { get; set; }
-        public double Dexterity { get; set; }
-        public double Constitution { get; set; }
-        public double Intelligence { get; set; }
-        public double Wisdom { get; set; }
-        public double Charisma { get; set; }
+        public int Strength { get; set; }
+        public int Dexterity { get; set; }
+        public int Constitution { get; set; }
+        public int Intelligence { get; set; }
+        public int Wisdom { get; set; }
+        public int Charisma { get; set; }
     }
 }
